refactor(WorkService.workMove): workMove 메서드 리팩토링
- Work도메인 statusOrder를 Long에서 Integer로 변환
- workMove 메서드에 Dto가 아닌, 직접 값을 꺼내서 전달
- workMove 메서드내 로직 분리 및 잘못된 요청 처리

+++ BE/src/main/java/codesquad/todo/web/works/dto/WorkMoveResponse.java
@@ -5,10 +5,10 @@ import lombok.Getter;
 
 @Getter
 public class WorkMoveResponse {
-    private Long order;
+    private Integer order;
     private WorkStatus status;
 
-    public WorkMoveResponse(WorkStatus status, Long order) {
+    public WorkMoveResponse(WorkStatus status, Integer order) {
         this.status = status;
         this.order = order;
     }
